IconGrid: Add PaginatedIConGrid and change IconGrid for allow us to use pagination
Add a subclass of IconGrid with proper functions to work with
pagination.
Also, we change the implementation of IconGrid to use the dynamic
spacing calculated before allocation of parents which allow us
to use the PaginationIconGrid, since we need to know before
PaginatedIconGrid allocation the parent size to know how many
pages we need to allocate all the apps, and therefore,how many
height we need for the PaginatedIconGrid.