AppDisplay: Use new IconGrid implemetation in AllView and FrequentView
Create a PaginatedAlphabeticalView to allow AllView to be a
paginatedView and adapt the implementation to that.
On the other hand FrequentView adapts the new IconGrid implementation.
Create a PaginationScrollView to allow us control the allocation of
the IconGrid child and the scroll adjustment.
Hook up the allocation/size comunication between the FrequentView,
AllView and FolderView with AppDisplay with a ViewStackLayout that
updates all views.